move mark evaluator to the mark plugin
